Extract raw window scan contract
