+++ packages/column-live-view-engine/src/topic-ordered-window.ts
@@ -1,4 +1,4 @@
-import type { TopicRawOrderByPlan, TopicRawPredicateFilterPlan } from "./row-scan";
+import type { TopicRawOrderByPlan, TopicRawPredicateFilterPlan } from "./raw-window-scan";
 import {
   compareQueryValue,
   isRangePlanValue,
